Add links, separate concerns
