+++ client/src/styles/Diagram.scss
@@ -1,15 +1,26 @@
 #diagram {
   width: 100%;
   height: 100%;
-  border: 2px solid brown;
 
   g.documents {
-    stroke-width: 3;
-    stroke: #000;
-    fill: red;
+    stroke-width: 2;
+    stroke: black;
+    fill: white;
 
     circle {
-      r: 0.7%;
+      r: 1%;
+
+      &:hover {
+        r: 1.2%;
+      }
+
+      transition: 350ms ease;
+    }
+  }
+
+  g.links {
+    line {
+      stroke: black;
     }
   }
 }
